// src/app/core/models/dependentDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface DependentDto {
  id: string;
  tenantId: string;
  memberId: string;
  name: string;
  email: string;
  address: string;
  phone1: string;
  phone2: string;
  identificationNumber: string;
  dependentType: string;
  dateOfBirth: string;
  createdAt: string;
  updatedAt: string;
  createdBy: string;
  updatedBy: string;
}